requirements.txt
## --- Core API ---
fastapi
uvicorn
pydantic

## --- OCR & File Processing ---
pytesseract
pdf2image
PyPDF2
pillow
odfpy
striprtf
html2text
python-docx
python-pptx
xlrd==1.2.0

## --- Google Cloud Vision OCR ---
google-cloud-vision
google-auth
grpcio

## --- Utilities ---
psutil
chardet
